backend/requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# CORS and middleware
python-multipart==0.0.6

# LLM and AI
openai==1.10.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-community==0.0.13

# Vector Database
chromadb==0.4.22

# Document Processing
pypdf==4.0.1
pymupdf==1.23.8
python-docx==1.1.0

# OCR
pytesseract==0.3.10
Pillow==10.2.0

# Audio Processing
openai-whisper==20231117
pyttsx3==2.90
gTTS==2.5.0
pydub==0.25.1

# Embeddings
sentence-transformers==2.3.1

# File Generation
reportlab==4.0.9
openpyxl==3.1.2

# HTTP Requests
httpx==0.26.0
requests==2.31.0

# Database
sqlalchemy==2.0.25

# Utilities
python-dotenv==1.0.0
pyyaml==6.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyjwt==2.8.0

# Google APIs (optional, for calendar and gmail)
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.116.0

# Notion and Todoist (optional)
notion-client==2.2.1

# Monitoring and logging
psutil==5.9.8

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
